+++ CHANGELOG.md
@@ -2,5 +2,17 @@
 # CGView.js Changelog
 --------------------------------------------------------------------------------
 
+## v1.1.0 - 2021-09-29
+
+- Remove 'id' attribute from Contig (Use 'name' instead)
+- Duplicate Contig and LegendItem names will automically become unique by appending a number
+- Added contig.isMapContig property
+- Changed contig.hasSeq() to contig.hasSeq (i.e. property instead of function)
+- Refresh feature labels when feature start/stop changes
+- Default Annotation font changed from 'sans-serif' to 'monospace'
+- Changed events:
+  - bookmark-shortcut -> bookmarks-shortcut
+  - changed 'drag' event to 'zoom'
+
 ## v1.0.0 - 2021-08-24
 - Initial Release